Implement CRUD operations for social entities and enhance upload middleware

+++ backend/service/socialService.js
@@ -1,5 +1,8 @@
+import { FOLDER } from '../constants/store.js';
+import removeFileMiddleware from '../middlewares/removeFile.js';
 import uploadMiddleware from '../middlewares/upload.js';
 import socialRepository from '../repositories/socialRepository.js';
+import { CustomError } from '../utils/customError.js';
 
 const socialSercive = {
   getSocials: async () => {
@@ -12,13 +15,43 @@ const socialSercive = {
     return await socialRepository.findByName(name);
   },
   createSocial: async (socialData) => {
-    const iconUrl = await uploadMiddleware(socialData.icon, 'uploads/icons');
+    const iconUrl = await uploadMiddleware(socialData.icon, FOLDER.ICON);
     return await socialRepository.create({ ...socialData, icon: iconUrl });
   },
   updateSocial: async (id, socialData) => {
-    return await socialRepository.update(id, socialData);
+    const { platform, url, icon } = socialData;
+
+    const social = await socialSercive.getSocialById(id);
+    if (!social) {
+      throw new CustomError('Social not found', 404);
+    }
+
+    let newIconUrl = social.icon;
+
+    if (typeof icon === 'object') {
+      newIconUrl = await uploadMiddleware(icon, FOLDER.ICON);
+      if (social.icon && newIconUrl !== social.icon) {
+        removeFileMiddleware(social.icon);
+      }
+    }
+
+    social.platform = platform || social.platform;
+    social.url = url || social.url;
+    social.icon = newIconUrl;
+
+    return await socialRepository.update(id, social);
   },
   deleteSocial: async (id) => {
+    const social = await socialSercive.getSocialById(id);
+    if (!social) {
+      throw new CustomError('Social not found', 404);
+    }
+
+    const iconUrl = social.icon;
+
+    if (iconUrl) {
+      removeFileMiddleware(iconUrl);
+    }
     return await socialRepository.delete(id);
   },
 };
